am335x: miiphy register address change
This patch is added as an workaround to support TLK110 PHY. The register
offset value of this PHY is exceeding 0xFF and cannot be used with
"unsigned char" datatype in miiphy_read, miiphy_write and miiphy_register
functions. Datatype of register offset is changed to unsigned short instead
of unsigned char so that offset value greater then 0xFF can be used.

Signed-off-by: Chandan Nath <chandan.nath@ti.com>