am335x: Enable CONFIG_SPL_BOARD_INIT
Vaibhav Bedia root-caused the SPL hang that is seen sometimes
as a problem of calling code that requires the BSS to be cleared
(and thus must be called post relocation) being called before.
The spl_board_init() hook is where we want to make this call now,
and i2c_init has already been taken care of for us.

Signed-off-by: Tom Rini <trini@ti.com>